feat(event) rework the prototype of lv_event_cb It encapsulates all event related parameters into a single lv_event_t obejct.

+++ src/misc/lv_mem.c
@@ -421,7 +421,7 @@ LV_ATTRIBUTE_FAST_MEM void lv_memset(void * dst, uint8_t v, size_t len)
         }
     }
 
-    uint32_t v32 = v + (v << 8) + (v << 16) + (v << 24);
+    uint32_t v32 = (uint32_t)v + ((uint32_t)v << 8) + ((uint32_t)v << 16) + ((uint32_t)v << 24);
 
     uint32_t * d32 = (uint32_t *)d8;
